src/functions/EmailIngestionProcess/requirements.txt
msal==1.27.0
requests==2.31.0
PyPDF2==3.0.1
pymongo==4.6.1
PyYAML==6.0.1
tqdm==4.66.1
